refactored module files
migrated files to be contained within their own modules, to reduce clutter and make organization easier.
