+++ modules/db/__init__.py
@@ -1,0 +1,1 @@
+# Used for python module creation
