+++ modules/standard/stdfn.py
